+++ src/Metal.jl
@@ -68,6 +68,7 @@ include("mapreduce.jl")
 include("accumulate.jl")
 include("indexing.jl")
 include("random.jl")
+include("fft.jl")
 
 # KernelAbstractions
 include("MetalKernels.jl")
